+++ endpoints/input-method-result.php
@@ -2,7 +2,8 @@
     require_once '_includes/functions.php';
     header("content-type: text/xml");
     echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n";
-    $inputMethod = getIvrResponse("input-method.php", $_REQUEST["SearchType"], ["1", "2", "3"]);
+    $searchType = getSearchType('SearchType');
+    $inputMethod = getIvrResponse("input-method.php", $_REQUEST['Digits'], ["1", "2", "3"]);
 
 if ($inputMethod == "3" && has_setting('jft_option') && json_decode(setting('jft_option'))) {
     ?>
@@ -12,8 +13,6 @@ if ($inputMethod == "3" && has_setting('jft_option') && json_decode(setting('jft
     <?php
     exit();
 }
-
-    $searchType = $_REQUEST['SearchType'];
 if (has_setting('province_lookup') && json_decode(setting('province_lookup'))) {
     $action = "province-voice-input.php";
 } else {
@@ -24,11 +23,11 @@ if (has_setting('province_lookup') && json_decode(setting('province_lookup'))) {
 <?php
 if ($inputMethod == "1") { // city or county
     ?>
-    <Redirect method="GET"><?php echo $action; ?>?SearchType=<?php echo $searchType; ?>&amp;InputMethod=<?php echo $inputMethod; ?></Redirect>
+    <Redirect method="GET"><?php echo $action; ?>?SearchType=<?php echo $_REQUEST['SearchType']; ?>&amp;InputMethod=<?php echo $inputMethod; ?></Redirect>
     <?php
 } else { // zip
     ?>
-    <Redirect method="GET">zip-input.php?SearchType=<?php echo $searchType; ?>&amp;InputMethod=<?php echo $inputMethod; ?></Redirect>
+    <Redirect method="GET">zip-input.php?SearchType=<?php echo $_REQUEST['SearchType']; ?>&amp;InputMethod=<?php echo $inputMethod; ?></Redirect>
     <?php
 }
 ?>