+++ endpoints/zip-input.php
@@ -3,9 +3,9 @@
     header("content-type: text/xml");
     echo "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n";
     
-    $searchType = $_REQUEST['SearchType'];
+    $searchType = getSearchType('SearchType');
     
-if ($searchType == "1") {
+if ($searchType == SearchType::VOLUNTEERS) {
     $action = "helpline-search.php";
 } else {
     $action = "address-lookup.php";
@@ -13,7 +13,7 @@ if ($searchType == "1") {
 
     $enterWord = has_setting('speech_gathering') && json_decode(setting('speech_gathering')) ? word('please_enter_or_say_your_digit') : word('please_enter_your_digit');
 
-    $action .= "?SearchType=" . $searchType;
+    $action .= "?SearchType=" . $_REQUEST['SearchType'];
 ?>
 <Response>
     <Gather input="<?php echo getInputType() ?>" numDigits="<?php echo setting('postal_code_length')?>" timeout="10" speechTimeout="auto" action="<?php echo $action; ?>" method="GET">